requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
xgboost>=1.7.0
imbalanced-learn>=0.10.0
mlxtend>=0.22.0
matplotlib>=3.6.0
seaborn>=0.12.0
pyyaml>=6.0
joblib>=1.2.0
